Changes to how the job's receivers are stored
A job now instead of having a list with all the receivers that where
available upon the job's creation, it has a "dual key" group and query,
the group specifies what group of receivers(a user can have many groups
of receivers e.g. "spamListA","spamListB" , and the query specifies a
subset inside this specific group of receivers.This makes better sense
when the job is repeatable since newly added receivers to a group will
be also candinate receivers for the job's next execution.